packages/vibe-cardano/pyproject.toml
[project]
name = "vibe-cardano"
version = "0.1.0"
description = "Cardano-specific node implementation"
requires-python = ">=3.14"
dependencies = [
    "cbor2pure>=5.8.0",
    "cryptography>=44.0.0",
    "httpx>=0.28.1",
    "pyarrow>=19.0.0",
    "pycardano>=0.19.2",
    "rich>=14.3.3",
    "uplc>=1.3.2",
    "vibe-core",
    "websockets>=16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vibe"]

# ---------------------------------------------------------------------------
# Native VRF extension build (optional — CMake-based)
# ---------------------------------------------------------------------------
# The _vrf_native pybind11 extension wraps the IOG libsodium fork for
# ECVRF-ED25519-SHA512-Elligator2 (draft-03) used by Cardano Praos.
#
# To build the native extension:
#   1. Ensure the submodule is initialized: git submodule update --init vendor/libsodium-iog
#   2. Ensure autotools are installed: autoconf, automake, libtool
#   3. Install build deps: pip install pybind11 cmake
#   4. Build: cd packages/vibe-cardano && cmake -B build && cmake --build build
#   5. Copy the .so: cp build/_vrf_native*.so src/vibe/cardano/crypto/
#
# The main package installs fine without the extension — VRF operations
# will raise NotImplementedError until the extension is built.
#
# Future: migrate to scikit-build-core for seamless pip install with
# native extension compilation. For now, the CMake build is separate
# to keep the hatchling build simple and fast for development.
